+++ backend-rust/src/migrations.rs
@@ -246,14 +246,17 @@ pub enum SchemaVersion {
     ReindexAffectedAccounts,
     #[display("0032:Reindex ginhash")]
     ReindexGinHash,
+    #[display("0033:Index for partial contracts search")]
+    IndexPartialContractsSearch,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking (destructive) database schema versions
     /// have been introduced since this version.
-    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::BakerPeriodApyViews;
+    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion =
+        SchemaVersion::IndexPartialContractsSearch;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::ReindexGinHash;
+    const LATEST: SchemaVersion = SchemaVersion::IndexPartialContractsSearch;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -305,6 +308,7 @@ impl SchemaVersion {
             SchemaVersion::ReindexAccountAccountStatementEntryType => false,
             SchemaVersion::ReindexAffectedAccounts => false,
             SchemaVersion::ReindexGinHash => false,
+            SchemaVersion::IndexPartialContractsSearch => false,
         }
     }
 
@@ -347,6 +351,7 @@ impl SchemaVersion {
             SchemaVersion::ReindexAccountAccountStatementEntryType => false,
             SchemaVersion::ReindexAffectedAccounts => false,
             SchemaVersion::ReindexGinHash => false,
+            SchemaVersion::IndexPartialContractsSearch => false,
         }
     }
 
@@ -550,7 +555,15 @@ impl SchemaVersion {
                     .await?;
                 SchemaVersion::ReindexGinHash
             }
-            SchemaVersion::ReindexGinHash => unimplemented!(
+            SchemaVersion::ReindexGinHash => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0033_index_for_partial_contracts_search.sql"
+                    )))
+                    .await?;
+                SchemaVersion::IndexPartialContractsSearch
+            }
+            SchemaVersion::IndexPartialContractsSearch => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),